restructured the way the inner product works
The object "InnerProduct" now just contains the weak form, bcs etc.
Once a ControlSpace is created, that control space then asks the inner
product for the matrix and the riesz map. this is necessary, because
the choices made by the ControlSpace change the exact form of the inner
product. E.g. if one wants to use high order deformations, then the
inner product needs to know about this and the matrix + riesz map need
to be assembled on high order functions